created component files and added function to DataUser.jsx
