+++ client/src/Components/DashboardComponents/PostDeduction.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function PostDeduction() {
+
+}
+
+export default PostDeduction;
