+++ client/src/Components/DashboardComponents/PostEarning.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const PostEarning = () => {
+
+}
+
+export default PostEarning;
