+++ client/src/Components/DashboardComponents/SubmitEarning.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SubmitEarning() {
+
+}
+
+export default SubmitEarning;
